# swagger/api.yaml
swagger: '2.0'
info:
  description: The public API for my.move.mil. This is a work in-progress, and is not a final product. We will be continuously updating this site based on feedback.
  version: 0.0.1
  title: my.move.mil
  license:
    name: MIT
    url: https://github.com/transcom/mymove/blob/master/LICENSE.md
basePath: /api/v1
consumes:
  - application/json
produces:
  - application/json
definitions:
  Address:
    type: object
    properties:
      street_address_1:
        type: string
        example: 123 Main Ave
        title: Address Line 1
      street_address_2:
        type: string
        example: Apartment 9000
        x-nullable: true
        title: Address Line 2
      street_address_3:
        type: string
        example: Montmârtre
        x-nullable: true
        title: Address Line 3
      city:
        type: string
        example: Anytown
        title: City
      state:
        title: State
        type: string
        example: CA
      postal_code:
        type: string
        description: zip code, international allowed
        format: zip
        title: ZIP/Postal Code
        example: "'90210' or 'N15 3NL'"
      country:
        type: string
        title: Country
        x-nullable: true
        example: "United States"
        default: United States
    required:
      - street_address_1
      - city
      - postal_code
...
  GBLOC:
    type: string
    description: Office Codes for Transcom offices originating GBLs
    pattern: '^[A-Z]{4}$'
    example: LHNQ
  IndexShipments:
    type: array
    items:
      $ref: '#/definitions/Shipment'
  Move:
    type: object
    properties:
      selected_move_type:
        $ref: '#/definitions/SelectedMoveType'
      orders_id:
        type: string
        format: uuid
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
      status:
        $ref: '#/definitions/MoveStatus'
      locator:
        type: string
        example: "12432"
      cancel_reason:
        type: string
        example: "Change of orders"
        x-nullable: true
    required:
      - orders_id
      - locator
  MoveStatus:
    type: string
    title: Move status
    enum:
      - DRAFT
      - SUBMITTED
      - APPROVED
      - COMPLETED
      - CANCELED
    x-display-value:
      DRAFT: Draft
      SUBMITTED: Submitted
      APPROVED: Approved
      COMPLETED: Completed
      CANCELED: Canceled
  SelectedMoveType:
    type: string
    title: Selected Move Type
    enum:
      - HHG
      - PPM
      - COMBO
    x-nullable: true
    x-display-value:
      HHG: Household Goods Move
      PPM: Personal Procured Move
      COMBO: Both HHG and PPM
...
  ServiceMember:
    type: object
    properties:
      first_name:
        type: string
        example: John
        x-nullable: true
        title: First Name
      middle_name:
        type: string
        example: L.
        x-nullable: true
        title: Middle Name
      last_name:
        type: string
        example: Donut
        x-nullable: true
        title: Last Name
      suffix:
        type: string
        example: Jr.
        x-nullable: true
        title: Suffix
      telephone:
        type: string
        format: telephone
        pattern: '^[2-9]\d{2}-\d{3}-\d{4}$'
        example: 212-555-5555
        x-nullable: true
        title: Best Contact Phone
      secondary_telephone:
        type: string
        format: telephone
        pattern: '^[2-9]\d{2}-\d{3}-\d{4}$'
        example: 212-555-5555
        x-nullable: true
        title: Secondary Phone
      personal_email:
        type: string
        format: x-email
        pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
        example: john_bob@example.com
        x-nullable: true
        title: Personal Email Address
      phone_is_preferred:
        type: boolean
        x-nullable: true
        title: Telephone
      text_message_is_preferred:
        type: boolean
        x-nullable: true
        title: Text Message - Coming Soon!
      email_is_preferred:
        type: boolean
        x-nullable: true
        title: Email
  TrafficDistributionList:
    type: object
    properties:
      source_rate_area:
        type: string
        example: US34
      destination_region:
        type: string
        example: "5"
      code_of_service:
        type: string
        example: D
        enum:
          - D
          - "2"
    required:
      - source_rate_area
      - destination_region
      - code_of_service
  Shipment:
    type: object
    properties:
      id:
        type: string
        format: uuid
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
        readOnly: true
      traffic_distribution_list:
        $ref: '#/definitions/TrafficDistributionList'
        readOnly: true
      service_member:
        $ref: '#/definitions/ServiceMember'
        readOnly: true
      pickup_date:
        type: string
        format: date-time
        example: 2018-04-21T17:32:28Z
      delivery_date:
        type: string
        format: date-time
        example: 2018-04-27T17:32:28Z
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
      source_gbloc:
        $ref: '#/definitions/GBLOC'
        readOnly: true
      destination_gbloc:
        $ref: '#/definitions/GBLOC'
        readOnly: true
      market:
        $ref: '#/definitions/ShipmentMarket'
        readOnly: true
        x-nullable: true
      book_date:
        type: string
        format: date
        example: "2018-03-15"
      requested_pickup_date:
        type: string
        format: date-time
        example: 2018-04-19T17:32:28Z
      move:
        $ref: '#/definitions/Move'
        x-nullable: true
      status:
        $ref: '#/definitions/ShipmentStatus'
        readOnly: true
      estimated_pack_days:
        type: integer
        description: Estimated days packing
        minimum: 0
        example: 2
        x-nullable: true
      estimated_transit_days:
        type: integer
        description: Estimated days in transit
        minimum: 0
        example: 30
        x-nullable: true
      pickup_address:
        $ref: '#/definitions/Address'
        x-nullable: true
      has_secondary_pickup_address:
        type: boolean
        title: 'Do you have household goods at any other pickup location?'
      secondary_pickup_address:
        $ref: '#/definitions/Address'
        x-nullable: true
      has_delivery_address:
        type: boolean
        title: 'Do you know the delivery address at destination yet?'
      delivery_address:
        $ref: '#/definitions/Address'
        x-nullable: true
      has_partial_sit_delivery_address:
        type: boolean
      partial_sit_delivery_address:
        $ref: '#/definitions/Address'
        x-nullable: true
      weight_estimate:
        type: integer
        description: Estimated weight of the shipment in lbs
        minimum: 0
        example: 10000
        x-nullable: true
        x-formatting: weight
        title: Weight Estimate
      progear_weight_estimate:
        type: integer
        description: Estimated weight of the progear in lbs
        minimum: 0
        example: 500
        x-nullable: true
        x-formatting: weight
        title: Progear Weight Estimate
      spouse_progear_weight_estimate:
        type: integer
        description: Estimated weight of the spouse progear in lbs
        minimum: 0
        example: 200
        x-nullable: true
        x-formatting: weight
        title: Spouse Progear Weight Estimate
      actual_weight:
        type: integer
        description: Actual weight of the shipment in lbs
        minimum: 0
        example: 10000
        x-nullable: true
        x-formatting: weight
        title: Actual Weight
      pm_survey_planned_pack_date:
        type: string
        format: date
        example: "2018-04-26"
        title: "Planned pack date"
        x-nullable: true
      pm_survey_planned_pickup_date:
        type: string
        format: date
        example: "2018-04-26"
        title: "Planned pickup date"
        x-nullable: true
      pm_survey_planned_delivery_date:
        type: string
        format: date
        example: "2018-04-26"
        title: "Planned delivery date"
        x-nullable: true
      pm_survey_weight_estimate:
        type: integer
        minimum: 0
        example: 10000
        title: Weight
        x-nullable: true
        x-formatting: weight
      pm_survey_progear_weight_estimate:
        type: integer
        minimum: 0
        example: 10000
        title: Pro-Gear weight
        x-nullable: true
        x-formatting: weight
      pm_survey_spouse_progear_weight_estimate:
        type: integer
        minimum: 0
        example: 10000
        title: Spouse's pro-gear
        x-nullable: true
        x-formatting: weight
      pm_survey_notes:
        type: string
        example: This document is good to go!
        x-nullable: true
        title: Notes
      pm_survey_method:
        type: string
        title: Survey method
        enum:
          - IN_PERSON
          - PHONE
          - VIDEO
        x-display-value:
          IN_PERSON: In person
          PHONE: Phone
          VIDEO: Video
...
  ShipmentMarket:
    type: string
    description: One of the possible 'Markets' for a shipment
    example: dHHG
    enum:
      - dHHG
      - iHHG
      - iUB
    x-display-value:
      dHHG: Domestic HHG
      iHHG: International HHG
      iUB: International unaccompanied baggage
  ShipmentStatus:
    type: string
    description: The stages in the lifecycle of a Shipment
    example: IN_PROGRESS
    enum: &SHIPMENTSTATUS
      - AWARDED
      - ACCEPTED
      - APPROVED
      - IN_TRANSIT
      - DELIVERED
      - COMPLETED
      - CANCELED
    x-display-value:
      AWARDED: Awarded
      ACCEPTED: Accepted
      REJECTED: Rejected
      IN_PROGRESS: In Progress
      IN_TRANSIT: In Transit
      DELIVERED: Delivered
      COMPLETED: Completed
      CANCELED: Canceled
...
paths:
...
  /shipments:
    get:
      summary: Gets visible shipments
      description: Allows a user to retrieve a list of shipments. The shipments are restricted to those they have authorization to view
      operationId: indexShipments
      tags:
        - shipments
      parameters:
        - name: status
          in: query
          type: array
          items:
            type: string
            enum: *SHIPMENTSTATUS
          collectionFormat: csv
          description: Restrict the list to shipments with matching ShipmentStatus
        - name: order_by
          in: query
          type: string
          description: Sort the returned list according to this ordering
          enum:
            - PICKUP_DATE_ASC
            - PICKUP_DATE_DESC
            - DELIVERY_DATE_ASC
            - DELIVERY_DATE_DESC
        - name: limit
          in: query
          type: integer
          description: maximum number of records to return
          minimum: 1
          default: 25
        - name: offset
          in: query
          type: integer
          description: offset into records to select from. Actual offset is this value multiplied by limit.
          minimum: 1
          default: 1
      responses:
        200:
          description: returns an array of Shipments
          schema:
            $ref: '#/definitions/IndexShipments'
        400:
          description: invalid request
        401:
          description: must be authenticated to use this endpoint
        403:
          description: not authorized to list these shipments
        422:
          description: cannot process request with given information
        500:
          description: server error
...